Move DB code to backend package
